+++ samples/qa-a11y/src/views/RadioItem.js
@@ -3,7 +3,7 @@ import React from 'react';
 
 import Section from '../components/Section';
 
-import css from '../App/App.module.less';
+import appCss from '../App/App.module.less';
 
 const RadioItemView = () => (
 	<>
@@ -12,7 +12,7 @@ const RadioItemView = () => (
 			<RadioItem alt="Disabled" disabled>Text 1</RadioItem>
 		</Section>
 
-		<Section className={css.marginTop} title="Aria-labelled">
+		<Section className={appCss.marginTop} title="Aria-labelled">
 			<RadioItem alt="Aria-labelled" aria-label="This is a Label 0.">Text 0</RadioItem>
 			<RadioItem alt="Aria-labelled and Disabled" aria-label="This is a Label 1." disabled>Text 1</RadioItem>
 		</Section>